# openapi.json
{
    "openapi": "3.0.1",
    "info": {
        "title": "Librería API",
        "description": "Documentación de la API",
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://localhost:8080",
            "description": "Generated server url"
        }
    ],
    "paths": {
...
        "/mensajeria/send": {
            "post": {
                "tags": [
                    "mensajeria-controller"
                ],
                "operationId": "sendMessage",
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/MensajeCreateRequest"
                            }
                        }
                    },
                    "required": true
                },
                "responses": {
                    "200": {
                        "description": "OK",
                        "content": {
                            "*/*": {
                                "schema": {
                                    "$ref": "#/components/schemas/Mensajeria"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/libros/scan-barcode": {
            "post": {
                "tags": [
                    "libro-controller"
                ],
                "operationId": "scanBarcode",
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/ISBNRequest"
                            }
                        }
                    },
                    "required": true
                },
                "responses": {
                    "200": {
                        "description": "OK",
                        "content": {
                            "*/*": {
                                "schema": {
                                    "$ref": "#/components/schemas/LibroScanResponse"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/libros/from-isbn": {
            "post": {
                "tags": [
                    "libro-controller"
                ],
                "operationId": "createFromIsbn",
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/LibroFromIsbnRequest"
                            }
                        }
                    },
                    "required": true
                },
                "responses": {
                    "200": {
                        "description": "OK",
                        "content": {
                            "*/*": {
                                "schema": {
                                    "$ref": "#/components/schemas/Libro"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/libros/add-libro": {
            "post": {
                "tags": [
                    "libro-controller"
                ],
                "operationId": "scanBarcode_1",
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/LibroFromIsbnRequest"
                            }
                        }
                    },
                    "required": true
                },
                "responses": {
                    "200": {
                        "description": "OK",
                        "content": {
                            "*/*": {
                                "schema": {
                                    "$ref": "#/components/schemas/Libro"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/intercambios": {
            "post": {
                "tags": [
                    "intercambio-controller"
                ],
                "operationId": "crearSolicitud",
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/IntercambioCreateRequest"
                            }
                        }
                    },
                    "required": true
                },
                "responses": {
                    "200": {
                        "description": "OK",
                        "content": {
                            "*/*": {
                                "schema": {
                                    "$ref": "#/components/schemas/Intercambio"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/intercambios/{id}/rechazar": {
            "post": {
                "tags": [
                    "intercambio-controller"
                ],
                "operationId": "rechazarIntercambio",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "format": "int64"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "content": {
                            "*/*": {
                                "schema": {
                                    "$ref": "#/components/schemas/Intercambio"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/intercambios/{id}/aceptar": {
            "post": {
                "tags": [
                    "intercambio-controller"
                ],
                "operationId": "aceptarIntercambio",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "format": "int64"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "content": {
                            "*/*": {
                                "schema": {
                                    "$ref": "#/components/schemas/Intercambio"
                                }
                            }
                        }
                    }
                }
            }
        },
...
        "/mensajeria/user/{userId}": {
            "get": {
                "tags": [
                    "mensajeria-controller"
                ],
                "operationId": "getMessagesForUser",
                "parameters": [
                    {
                        "name": "userId",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "format": "int64"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "content": {
                            "*/*": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "$ref": "#/components/schemas/Mensajeria"
                                    }
                                }
                            }
                        }
                    }
                }
            }
        },
        "/mensajeria/conversation/{userA}/{userB}": {
            "get": {
                "tags": [
                    "mensajeria-controller"
                ],
                "operationId": "getConversation",
                "parameters": [
                    {
                        "name": "userA",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "format": "int64"
                        }
                    },
                    {
                        "name": "userB",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "format": "int64"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "content": {
                            "*/*": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "$ref": "#/components/schemas/Mensajeria"
                                    }
                                }
                            }
                        }
                    }
                }
            }
        },
        "/intercambios/solicitante/{id}": {
            "get": {
                "tags": [
                    "intercambio-controller"
                ],
                "operationId": "porSolicitante",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "format": "int64"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "content": {
                            "*/*": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "$ref": "#/components/schemas/Intercambio"
                                    }
                                }
                            }
                        }
                    }
                }
            }
        },
        "/intercambios/libro/{id}": {
            "get": {
                "tags": [
                    "intercambio-controller"
                ],
                "operationId": "historialPorLibro",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "format": "int64"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "content": {
                            "*/*": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "$ref": "#/components/schemas/Intercambio"
                                    }
                                }
                            }
                        }
                    }
                }
            }
        },
        "/intercambios/historial/{usuarioId1}/{usuarioId2}": {
            "get": {
                "tags": [
                    "intercambio-controller"
                ],
                "operationId": "historialEntre",
                "parameters": [
                    {
                        "name": "usuarioId1",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "format": "int64"
                        }
                    },
                    {
                        "name": "usuarioId2",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "format": "int64"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "content": {
                            "*/*": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "$ref": "#/components/schemas/Intercambio"
                                    }
                                }
                            }
                        }
                    }
                }
            }
        },
        "/intercambios/destinatario/{id}": {
            "get": {
                "tags": [
                    "intercambio-controller"
                ],
                "operationId": "porDestinatario",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "format": "int64"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "content": {
                            "*/*": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "$ref": "#/components/schemas/Intercambio"
                                    }
                                }
                            }
                        }
                    }
                }
            }
        },
...
    },
    "components": {
        "schemas": {
            "Autor": {
                "type": "object",
                "properties": {
                    "id": {
                        "type": "integer",
                        "format": "int64"
                    },
                    "nombre": {
                        "type": "string"
                    }
                }
            },
            "Genero": {
                "type": "object",
                "properties": {
                    "id": {
                        "type": "integer",
                        "format": "int64"
                    },
                    "nombre": {
                        "type": "string"
                    }
                }
            },
            "Libro": {
                "type": "object",
                "properties": {
                    "codigoBarra": {
                        "type": "string"
                    },
                    "titulo": {
                        "type": "string"
                    },
                    "descripcion": {
                        "type": "string"
                    },
                    "anio_publicacion": {
                        "type": "string",
                        "format": "date-time"
                    },
                    "portada_url": {
                        "type": "string"
                    },
                    "status": {
                        "type": "string"
                    },
                    "propietario": {
                        "$ref": "#/components/schemas/Usuario"
                    },
                    "autores": {
                        "uniqueItems": true,
                        "type": "array",
                        "items": {
                            "$ref": "#/components/schemas/Autor"
                        }
                    },
                    "generos": {
                        "uniqueItems": true,
                        "type": "array",
                        "items": {
                            "$ref": "#/components/schemas/Genero"
                        }
                    }
                }
            },
...
            "Usuario": {
                "type": "object",
                "properties": {
                    "id": {
                        "type": "integer",
                        "format": "int64"
                    },
                    "nombre": {
                        "type": "string"
                    },
                    "apellido": {
                        "type": "string"
                    },
                    "correo": {
                        "type": "string"
                    },
                    "contrasena": {
                        "type": "string"
                    }
                }
            },
...
            "MensajeCreateRequest": {
                "type": "object",
                "properties": {
                    "senderId": {
                        "type": "integer",
                        "format": "int64"
                    },
                    "recipientId": {
                        "type": "integer",
                        "format": "int64"
                    },
                    "contenido": {
                        "type": "string"
                    }
                }
            },
            "Mensajeria": {
                "type": "object",
                "properties": {
                    "id": {
                        "type": "integer",
                        "format": "int64"
                    },
                    "senderId": {
                        "type": "integer",
                        "format": "int64"
                    },
                    "recipientId": {
                        "type": "integer",
                        "format": "int64"
                    },
                    "asunto": {
                        "type": "string"
                    },
                    "destinatario": {
                        "type": "string"
                    },
                    "cuerpo_mensaje": {
                        "type": "string"
                    },
                    "timestamp": {
                        "type": "string",
                        "format": "date-time"
                    },
                    "leido": {
                        "type": "boolean"
                    }
                }
            },
            "ISBNRequest": {
                "type": "object",
                "properties": {
                    "isbn": {
                        "type": "string"
                    }
                }
            },
            "LibroScanResponse": {
                "type": "object",
                "properties": {
                    "isbn": {
                        "type": "string"
                    },
                    "titulo": {
                        "type": "string"
                    },
                    "autores": {
                        "type": "array",
                        "items": {
                            "type": "string"
                        }
                    },
                    "generos": {
                        "type": "array",
                        "items": {
                            "type": "string"
                        }
                    },
                    "portadaUrl": {
                        "type": "string"
                    }
                }
            },
            "LibroFromIsbnRequest": {
                "type": "object",
                "properties": {
                    "isbn": {
                        "type": "string"
                    }
                }
            },
            "IntercambioCreateRequest": {
                "type": "object",
                "properties": {
                    "solicitanteId": {
                        "type": "integer",
                        "format": "int64"
                    },
                    "libroId": {
                        "type": "integer",
                        "format": "int64"
                    },
                    "destinatarioId": {
                        "type": "integer",
                        "format": "int64"
                    }
                }
            },
            "Intercambio": {
                "type": "object",
                "properties": {
                    "id": {
                        "type": "integer",
                        "format": "int64"
                    },
                    "libro": {
                        "$ref": "#/components/schemas/Libro"
                    },
                    "solicitante": {
                        "$ref": "#/components/schemas/Usuario"
                    },
                    "destinatario": {
                        "$ref": "#/components/schemas/Usuario"
                    },
                    "estado": {
                        "type": "string",
                        "enum": [
                            "SOLICITADO",
                            "ACEPTADO",
                            "RECHAZADO",
                            "EN_PROGRESO",
                            "FINALIZADO",
                            "CANCELADO"
                        ]
                    },
                    "fechaSolicitud": {
                        "type": "string",
                        "format": "date-time"
                    },
                    "fechaAceptacion": {
                        "type": "string",
                        "format": "date-time"
                    },
                    "fechaFinalizacion": {
                        "type": "string",
                        "format": "date-time"
                    }
                }
            },
...
        }
    }
}