# src/main/java/com/example/libreria_app/dto/ISBNRequest.java
package com.example.libreria_app.dto;

public class ISBNRequest {
    private String isbn;

    public ISBNRequest() {
    }

    public String getIsbn() {
        return isbn;
    }

    public void setIsbn(String isbn) {
        this.isbn = isbn;
    }
}