# src/main/java/com/example/libreria_app/dto/LibroFromIsbnRequest.java
package com.example.libreria_app.dto;

public class LibroFromIsbnRequest {
    private String isbn;

    public LibroFromIsbnRequest() {}

    public String getIsbn() {
        return isbn;
    }

    public void setIsbn(String isbn) {
        this.isbn = isbn;
    }
}