# src/main/java/com/example/libreria_app/dto/LibroScanResponse.java
package com.example.libreria_app.dto;

import java.util.List;

public class LibroScanResponse {
    private String isbn;
    private String titulo;
    private List<String> autores;
    private List<String> generos;
    private String portadaUrl;

    public LibroScanResponse() {
    }

    public String getIsbn() {
        return isbn;
    }

    public void setIsbn(String isbn) {
        this.isbn = isbn;
    }

    public String getTitulo() {
        return titulo;
    }

    public void setTitulo(String titulo) {
        this.titulo = titulo;
    }

    public List<String> getAutores() {
        return autores;
    }

    public void setAutores(List<String> autores) {
        this.autores = autores;
    }

    public List<String> getGeneros() {
        return generos;
    }

    public void setGeneros(List<String> generos) {
        this.generos = generos;
    }

    public String getPortadaUrl() {
        return portadaUrl;
    }

    public void setPortadaUrl(String portadaUrl) {
        this.portadaUrl = portadaUrl;
    }
    
}