# src/main/java/com/example/libreria_app/dto/MensajeCreateRequest.java
package com.example.libreria_app.dto;

public class MensajeCreateRequest {
    private Long senderId;
    private Long recipientId;
    private String contenido;

    public MensajeCreateRequest() {
    }

    public Long getSenderId() {
        return senderId;
    }

    public void setSenderId(Long senderId) {
        this.senderId = senderId;
    }

    public Long getRecipientId() {
        return recipientId;
    }

    public void setRecipientId(Long recipientId) {
        this.recipientId = recipientId;
    }

    public String getContenido() {
        return contenido;
    }

    public void setContenido(String contenido) {
        this.contenido = contenido;
    }
}